requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bootstrap4==0.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.5
crispy-bootstrap4==2024.1
Django==4.2.3
django-ckeditor==6.7.1
django-crispy-forms==2.0
django-database-url==1.0.3
django-dotenv==1.4.2
django-js-asset==2.2.0
filelock==3.14.0
idna==3.7
jinja2==3.1.4
joblib==1.4.2
lit==18.1.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pillow==10.3.0
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
python-dotenv==1.0.1
regex==2024.5.15
requests==2.32.3
sqlparse==0.4.4
stripe==9.10.0
sympy==1.12.1
torch==2.0.1
tqdm==4.66.4
triton==2.0.0
typing-extensions==4.12.2
tzdata==2023.3
urllib3==2.2.1
